Move HttpServerManager to common
Move HttpServerManager class to common
Fix mypy errors
